moved contributed regularization and cost functions to math namespace
