+++ numl/Math/Functions/Cost/ICostFunction.cs
@@ -1,11 +1,10 @@
-﻿using numl.Math.Functions;
-using numl.Math.LinearAlgebra;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Linq;
-using System.Text;
+using numl.Math.LinearAlgebra;
+using System.Collections.Generic;
+using numl.Math.Functions.Regularization;
 
-namespace numl.Functions
+namespace numl.Math.Functions.Cost
 {
     /// <summary>
     /// Cost function interface
