+++ numl/Math/Functions/Cost/LinearCostFunction.cs
@@ -1,16 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using numl.Math.LinearAlgebra;
-using numl.Math.Functions;
+using System.Collections.Generic;
+using numl.Math.Functions.Regularization;
 
-namespace numl.Functions.CostFunctions
+namespace numl.Math.Functions.Cost
 {
     /// <summary>
     /// A Linear Cost Function.
     /// </summary>
-    public class LinearCostFunction : ICostFunction
+    public class LinearCostFunction : numl.Math.Functions.Cost.ICostFunction
     {
                 /// <summary>
         /// Compute the error cost of the given Theta parameter for the training and label sets
