+++ numl/Math/Functions/Regularization/IRegularizer.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace numl.Functions
+namespace numl.Math.Functions.Regularization
 {
     /// <summary>
     /// Regularization function
