+++ numl/Math/Optimization/GradientDescent.cs
@@ -5,7 +5,7 @@ using System.Text;
 using numl.Math.LinearAlgebra;
 using numl.Math.Functions;
 
-namespace numl.Functions
+namespace numl.Math.Optimization
 {
     /// <summary>
     /// Gradient Descent
@@ -25,7 +25,7 @@ namespace numl.Functions
         /// <param name="regularizer">The regularization function to apply</param>
         /// <returns></returns>
         public static Tuple<double, Vector> Run(Vector theta, Matrix x, Vector y, int maxIterations, double learningRateAlpha, 
-            ICostFunction costFunction, double lambda, IRegularizer regularizer)
+            numl.Math.Functions.Cost.ICostFunction costFunction, double lambda, numl.Math.Functions.Regularization.IRegularizer regularizer)
         {
             Vector bestTheta = theta.Copy();
             double bestCost = double.PositiveInfinity;
